Split out opsd file
