+++ swb_ecco/data_ingestion/__init__.py
@@ -1,0 +1,5 @@
+from swb_ecco.data_ingestion.opsd import DataLoaderOPSD
+
+__all__ = [
+    'DataLoaderOPSD'
+]
